move findBlogPost to utils
