+++ src/pages/plage-des-catalans.tsx
@@ -1,5 +1,6 @@
 import ArticleLayout from '../Components/ArticleLayout';
-import { findBlogPost, ArticleMetadata } from '../types';
+import { ArticleMetadata } from '../types';
+import { findBlogPost } from '../utils/blog';
 
 const Page = () => {
     const blog = findBlogPost(b => b.title.toLowerCase() === "plage des catalans");
