+++ src/types.ts
@@ -1,5 +1,4 @@
 import { ReactNode } from 'react';
-import blogPostArchive from './blogPostArchive.json';
 
 export interface BlogPost {
   title: string;
@@ -28,13 +27,4 @@ export interface ArticleLayoutProps {
 
 export interface YouTubeEmbedProps {
   videoId: string;
-}
-
-export function findBlogPost(matcher: (blog: BlogPost) => boolean): BlogPost {
-  const posts: BlogPost[] = blogPostArchive;
-  const blog = posts.find(matcher);
-  if (!blog) {
-    throw new Error('Blog post not found');
-  }
-  return blog;
 }
